linebot/pyproject.toml
[tool.poetry]
name = "ncku-chatbot"
version = "0.1.0"
description = ""
authors = ["yenslife <77geo5rge6@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
line-bot-sdk = "^3.14.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
openai = "^1.70.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
